refacto getsearchresultbysearchquery to queries and remove pg transaction
