+++ back/src/domains/establishment/adapters/PgEstablishmentAggregateRepository.integration.test.ts
@@ -60,7 +60,6 @@ import {
   establishmentWithFitForDisabledWorkersYesCertified,
   establishmentWithFitForDisabledWorkersYesDeclaredOnly,
   establishmentWithOfferA1101_AtPosition,
-  establishmentWithOfferA1101_AtPositionWithTwoLocations,
   establishmentWithOfferA1101_close,
   establishmentWithOfferA1101_outOfDistanceRange,
   groomChevauxOffer,
@@ -2280,1496 +2279,6 @@ describe("PgEstablishmentAggregateRepository", () => {
         });
       });
     });
-    describe("legacySearchImmersionResults", () => {
-      describe("with `maxResults` parameter", () => {
-        beforeEach(async () => {
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            establishmentWithOfferA1101_AtPosition,
-          );
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            establishmentWithOfferA1101_close,
-          );
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            establishmentWithOfferA1101_outOfDistanceRange,
-          );
-        });
-
-        it("returns 1 closest establishment with `maxResults` at 1", async () => {
-          expectToEqual(
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: searchMadeDistanceWithoutRome,
-                maxResults: 1,
-              },
-            ),
-            [
-              makeExpectedSearchResult({
-                establishment: establishmentWithOfferA1101_AtPosition,
-                withOffers: [offer_A1101_11987],
-                withLocationAndDistance: {
-                  ...locationOfSearchPosition,
-                  distance: 0,
-                },
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-            ],
-          );
-        });
-
-        it("returns 2 closest establishments with `maxResults` at 2", async () => {
-          expectToEqual(
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: searchMadeDistanceWithoutRome,
-                maxResults: 2,
-              },
-            ),
-            [
-              makeExpectedSearchResult({
-                establishment: establishmentWithOfferA1101_AtPosition,
-                withOffers: [offer_A1101_11987],
-                withLocationAndDistance: {
-                  ...locationOfSearchPosition,
-                  distance: 0,
-                },
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-              makeExpectedSearchResult({
-                establishment: establishmentWithOfferA1101_close,
-                withOffers: [offer_A1101_11987],
-                withLocationAndDistance: {
-                  ...locationOfCloseSearchPosition,
-                  distance: 133.12254555,
-                },
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-            ],
-          );
-        });
-
-        it("returns 2 closest establishments with `maxResults` at 3 without too far establishment", async () => {
-          expectArraysToMatch(
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: searchMadeDistanceWithoutRome,
-                maxResults: 3,
-              },
-            ),
-            [
-              makeExpectedSearchResult({
-                establishment: establishmentWithOfferA1101_AtPosition,
-                withOffers: [offer_A1101_11987],
-                withLocationAndDistance: {
-                  ...locationOfSearchPosition,
-                  distance: 0,
-                },
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-              makeExpectedSearchResult({
-                establishment: establishmentWithOfferA1101_close,
-                withOffers: [offer_A1101_11987],
-                withLocationAndDistance: {
-                  ...locationOfCloseSearchPosition,
-                  distance: 133.12254555,
-                },
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-            ],
-          );
-        });
-        it("should always return the same number of results regardless of the sortedBy parameter", async () => {
-          await kyselyDb.deleteFrom("establishments__users").execute();
-          await kyselyDb.deleteFrom("immersion_offers").execute();
-          await kyselyDb.deleteFrom("discussions").execute();
-          await kyselyDb.deleteFrom("establishments_location_infos").execute();
-          await kyselyDb
-            .deleteFrom("establishments_location_positions")
-            .execute();
-          await kyselyDb.deleteFrom("establishments").execute();
-          await kyselyDb.deleteFrom("banned_establishments").execute();
-
-          const INDEX_OF_NOT_AVAILABLE_ESTABLISHMENT = 2;
-          const INDEX_OF_CURRENTLY_UNAVAILABLE_ESTABLISHMENT = 3;
-          const establishmentAggregatesByDateAndScoreDescending: EstablishmentAggregate[] =
-            arrayFromNumber(10).map((i) => ({
-              ...searchableByAllEstablishment,
-              establishment: {
-                ...searchableByAllEstablishment.establishment,
-                siret: `9999999999999${i}`,
-                updatedAt: addDays(new Date(), -i),
-                isMaxDiscussionsForPeriodReached:
-                  i === INDEX_OF_NOT_AVAILABLE_ESTABLISHMENT,
-                score: 10 - i,
-                nextAvailabilityDate:
-                  i === INDEX_OF_CURRENTLY_UNAVAILABLE_ESTABLISHMENT
-                    ? addDays(new Date(), i).toISOString()
-                    : undefined,
-                locations: [
-                  {
-                    ...searchableByAllEstablishment.establishment.locations[0],
-                    id: uuid(),
-                  },
-                ],
-              },
-            }));
-
-          const displayedResults = [0, 1, 4, 5, 6];
-
-          const expectedResults = displayedResults.map((i) =>
-            makeExpectedSearchResult({
-              establishment: establishmentAggregatesByDateAndScoreDescending[i],
-              withOffers:
-                establishmentAggregatesByDateAndScoreDescending[i].offers,
-              withLocationAndDistance: {
-                ...establishmentAggregatesByDateAndScoreDescending[i]
-                  .establishment.locations[0],
-                distance: 0,
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          );
-
-          await Promise.all(
-            establishmentAggregatesByDateAndScoreDescending.map(
-              (establishmentAggregate) =>
-                pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-                  establishmentAggregate,
-                ),
-            ),
-          );
-
-          const sortedByDateResults =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  ...searchMadeDistanceWithoutRome,
-                  sortedBy: "date",
-                },
-                maxResults: 5,
-              },
-            );
-
-          expectArraysToEqual(sortedByDateResults, expectedResults);
-
-          const sortedByScoreResults =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  ...searchMadeDistanceWithoutRome,
-                  sortedBy: "score",
-                },
-                maxResults: 5,
-              },
-            );
-          expectArraysToEqual(sortedByScoreResults, expectedResults);
-
-          const sortedByDistanceResults =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  ...searchMadeDistanceWithoutRome,
-                  sortedBy: "distance",
-                },
-                maxResults: 5,
-              },
-            );
-          expectArraysToEqual(sortedByDistanceResults, expectedResults);
-        });
-      });
-
-      describe("SearchMade parameters", () => {
-        describe("'establishmentSearchableBy' SearchMade parameter", () => {
-          beforeEach(async () => {
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              searchableByAllEstablishment,
-            );
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              searchableByStudentsEstablishment,
-            );
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              searchableByJobSeekerEstablishment,
-            );
-          });
-
-          it(`with "establishmentSearchableBy:'students'" return only establishments SearchImmersionResults searchable by student`, async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: {
-                    ...cartographeSearchMade,
-                    searchableBy: "students",
-                  },
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [searchableByStudentsEstablishment, searchableByAllEstablishment]
-                .map((establishment) =>
-                  makeExpectedSearchResult({
-                    establishment: establishment,
-                    withLocationAndDistance: {
-                      ...establishment.establishment.locations[0],
-                      distance: 0,
-                    },
-                    withOffers: establishment.offers,
-                    nafLabel: "Activités des agences de travail temporaire",
-                  }),
-                )
-                .sort(
-                  sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-                ),
-            );
-          });
-
-          it(`with "establishmentSearchableBy:'jobSeekers'" return only establishments SearchImmersionResults searchable by jobseekers`, async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: {
-                    ...cartographeSearchMade,
-                    searchableBy: "jobSeekers",
-                  },
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [searchableByJobSeekerEstablishment, searchableByAllEstablishment]
-                .map((establishment) =>
-                  makeExpectedSearchResult({
-                    establishment: establishment,
-                    withLocationAndDistance: {
-                      ...establishment.establishment.locations[0],
-                      distance: 0,
-                    },
-                    withOffers: establishment.offers,
-                    nafLabel: "Activités des agences de travail temporaire",
-                  }),
-                )
-                .sort(
-                  sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-                ),
-            );
-          });
-
-          it('with"establishmentSearchableBy" not defined return all establishment SearchImmersionResults', async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: cartographeSearchMade,
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [
-                searchableByStudentsEstablishment,
-                searchableByJobSeekerEstablishment,
-                searchableByAllEstablishment,
-              ]
-                .map((establishment) =>
-                  makeExpectedSearchResult({
-                    establishment: establishment,
-                    withLocationAndDistance: {
-                      ...establishment.establishment.locations[0],
-                      distance: 0,
-                    },
-                    withOffers: establishment.offers,
-                    nafLabel: "Activités des agences de travail temporaire",
-                  }),
-                )
-                .sort(
-                  sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-                ),
-            );
-          });
-        });
-
-        describe("without 'romeCode' SearchMade parameter", () => {
-          it("returns all establishments within geographical area", async () => {
-            // Prepare
-            const establishmentCloseWithA1101Offers =
-              new EstablishmentAggregateBuilder(
-                establishmentWithOfferA1101_AtPosition,
-              )
-                .withOffers([offer_A1101_20404, offer_A1101_17751])
-                .build();
-            const establishmentFarWithA1101Offers =
-              new EstablishmentAggregateBuilder(
-                establishmentWithOfferA1101_outOfDistanceRange,
-              )
-                .withOffers([offer_A1101_12862])
-                .build();
-
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishmentCloseWithA1101Offers,
-            );
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishmentFarWithA1101Offers,
-            );
-
-            // Act & Assert
-            expectToEqual(
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: searchMadeDistanceWithoutRome,
-                },
-              ),
-              [
-                makeExpectedSearchResult({
-                  establishment: establishmentCloseWithA1101Offers,
-                  withOffers: [offer_A1101_17751, offer_A1101_20404],
-                  withLocationAndDistance: {
-                    ...locationOfSearchPosition,
-                    distance: 0,
-                  },
-                  nafLabel: "Activités des agences de travail temporaire",
-                }),
-              ],
-            );
-          });
-
-          it("returns only offers with locations within geographical area without rome code given", async () => {
-            const establishmentAggregateAtSaintesAndVeaux =
-              new EstablishmentAggregateBuilder()
-                .withEstablishmentSiret("78000403200029")
-                .withOffers([cuvisteOffer, artisteCirqueOffer])
-                .withLocations([
-                  bassompierreSaintesLocation,
-                  veauxLocation, // outside geographical area
-                ])
-                .withUserRights([osefUserRight])
-                .build();
-
-            const establishmentAggregateAtChaniersAndLaRochelle =
-              new EstablishmentAggregateBuilder()
-                .withEstablishmentSiret("78000403200030")
-                .withOffers([
-                  cartographeImmersionOffer,
-                  cuvisteOffer,
-                  groomChevauxOffer,
-                ])
-                .withLocations([
-                  portHubleChaniersLocation,
-                  tourDeLaChaineLaRochelleLocation,
-                ])
-                .withUserRights([osefUserRight])
-                .build();
-
-            // Prepare
-            /// Two establishments located inside geographical area
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishmentAggregateAtSaintesAndVeaux,
-            );
-
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishmentAggregateAtChaniersAndLaRochelle,
-            );
-
-            // Act
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: {
-                    sortedBy: "date",
-                    distanceKm: 100,
-                    ...centerOfSaintesGeoPosition,
-                  },
-                },
-              );
-
-            // Assert
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [
-                ...[cuvisteOffer, artisteCirqueOffer].map((offer) =>
-                  makeExpectedSearchResult({
-                    establishment: establishmentAggregateAtSaintesAndVeaux,
-                    withOffers: [offer],
-                    withLocationAndDistance: {
-                      ...bassompierreSaintesLocation,
-                      distance: 7704.55035665,
-                    },
-                    nafLabel: "Activités des agences de travail temporaire",
-                  }),
-                ),
-                ...[
-                  cartographeImmersionOffer,
-                  cuvisteOffer,
-                  groomChevauxOffer,
-                ].map((offer) =>
-                  makeExpectedSearchResult({
-                    establishment:
-                      establishmentAggregateAtChaniersAndLaRochelle,
-                    withOffers: [offer],
-                    withLocationAndDistance: {
-                      ...portHubleChaniersLocation,
-                      distance: 11093.36505388,
-                    },
-                    nafLabel: "Activités des agences de travail temporaire",
-                  }),
-                ),
-                ...[
-                  cartographeImmersionOffer,
-                  cuvisteOffer,
-                  groomChevauxOffer,
-                ].map((offer) =>
-                  makeExpectedSearchResult({
-                    establishment:
-                      establishmentAggregateAtChaniersAndLaRochelle,
-                    withOffers: [offer],
-                    withLocationAndDistance: {
-                      ...tourDeLaChaineLaRochelleLocation,
-                      distance: 56222.51061222,
-                    },
-                    nafLabel: "Activités des agences de travail temporaire",
-                  }),
-                ),
-              ].sort(sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults),
-            );
-          });
-        });
-
-        describe("'nafCodes' SearchMade param", () => {
-          beforeEach(async () => {
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishment0145Z_A,
-            );
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishment0145Z_B,
-            );
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishment9900Z,
-            );
-            await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishment4741Z,
-            );
-          });
-
-          it("establishments with all naf when naf filter is not provided", async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: { sortedBy: "date" },
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [
-                makeExpectedSearchResult({
-                  establishment: establishment0145Z_A,
-                  withLocationAndDistance: {
-                    ...establishment0145Z_A.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment0145Z_A.offers,
-                  nafLabel: "Élevage d'ovins et de caprins",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment0145Z_B,
-                  withLocationAndDistance: {
-                    ...establishment0145Z_B.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment0145Z_B.offers,
-                  nafLabel: "Élevage d'ovins et de caprins",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment4741Z,
-                  withLocationAndDistance: {
-                    ...establishment4741Z.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment4741Z.offers,
-                  nafLabel:
-                    "Commerce de détail d'ordinateurs, d'unités périphériques et de logiciels en magasin spécialisé",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment9900Z,
-                  withLocationAndDistance: {
-                    ...establishment9900Z.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment9900Z.offers,
-                  nafLabel:
-                    "Activités des organisations et organismes extraterritoriaux",
-                }),
-              ].sort(sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults),
-            );
-          });
-
-          it("establishments with all naf when naf filter is empty", async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: { sortedBy: "date", nafCodes: [] },
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [
-                makeExpectedSearchResult({
-                  establishment: establishment0145Z_A,
-                  withLocationAndDistance: {
-                    ...establishment0145Z_A.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment0145Z_A.offers,
-                  nafLabel: "Élevage d'ovins et de caprins",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment0145Z_B,
-                  withLocationAndDistance: {
-                    ...establishment0145Z_B.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment0145Z_B.offers,
-                  nafLabel: "Élevage d'ovins et de caprins",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment4741Z,
-                  withLocationAndDistance: {
-                    ...establishment4741Z.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment4741Z.offers,
-                  nafLabel:
-                    "Commerce de détail d'ordinateurs, d'unités périphériques et de logiciels en magasin spécialisé",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment9900Z,
-                  withLocationAndDistance: {
-                    ...establishment9900Z.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment9900Z.offers,
-                  nafLabel:
-                    "Activités des organisations et organismes extraterritoriaux",
-                }),
-              ].sort(sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults),
-            );
-          });
-
-          it("establishments with naf when naf filter is provided with one value", async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: {
-                    sortedBy: "date",
-                    nafCodes: [establishment0145Z_A.establishment.nafDto.code],
-                  },
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [establishment0145Z_A, establishment0145Z_B]
-                .map((establishment) =>
-                  makeExpectedSearchResult({
-                    establishment: establishment,
-                    withLocationAndDistance: {
-                      ...establishment.establishment.locations[0],
-                      distance: undefined,
-                    },
-                    withOffers: establishment.offers,
-                    nafLabel: "Élevage d'ovins et de caprins",
-                  }),
-                )
-                .sort(
-                  sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-                ),
-            );
-          });
-
-          it("establishments with different naf when naf filter is provided with multiple values", async () => {
-            const results =
-              await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-                {
-                  searchMade: {
-                    sortedBy: "date",
-                    nafCodes: [
-                      establishment0145Z_A.establishment.nafDto.code,
-                      establishment4741Z.establishment.nafDto.code,
-                    ],
-                  },
-                },
-              );
-
-            expectToEqual(
-              results.sort(
-                sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults,
-              ),
-              [
-                makeExpectedSearchResult({
-                  establishment: establishment0145Z_A,
-                  withLocationAndDistance: {
-                    ...establishment0145Z_A.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment0145Z_A.offers,
-                  nafLabel: "Élevage d'ovins et de caprins",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment0145Z_B,
-                  withLocationAndDistance: {
-                    ...establishment0145Z_B.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment0145Z_B.offers,
-                  nafLabel: "Élevage d'ovins et de caprins",
-                }),
-                makeExpectedSearchResult({
-                  establishment: establishment4741Z,
-                  withLocationAndDistance: {
-                    ...establishment4741Z.establishment.locations[0],
-                    distance: undefined,
-                  },
-                  withOffers: establishment4741Z.offers,
-                  nafLabel:
-                    "Commerce de détail d'ordinateurs, d'unités périphériques et de logiciels en magasin spécialisé",
-                }),
-              ].sort(sortSearchResultsByDistanceAndRomeAndSiretOnRandomResults),
-            );
-          });
-        });
-      });
-
-      describe("only available offers", () => {
-        it("doesn't show unavailable establishement (legacy behavior)", async () => {
-          const establishmentsTestSet = [
-            searchableByAllEstablishment,
-            searchableByJobSeekerEstablishment,
-            unavailableEstablishment,
-          ];
-          await Promise.all(
-            establishmentsTestSet.map((establishmentAgg) =>
-              pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-                establishmentAgg,
-              ),
-            ),
-          );
-          const results =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: { sortedBy: "date" },
-              },
-            );
-          expectToEqual(results, [
-            makeExpectedSearchResult({
-              establishment: searchableByAllEstablishment,
-              withOffers: searchableByAllEstablishment.offers,
-              withLocationAndDistance: {
-                ...searchableByAllEstablishment.establishment.locations[0],
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-            makeExpectedSearchResult({
-              establishment: searchableByJobSeekerEstablishment,
-              withOffers: searchableByJobSeekerEstablishment.offers,
-              withLocationAndDistance: {
-                ...searchableByJobSeekerEstablishment.establishment
-                  .locations[0],
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ]);
-        });
-      });
-
-      describe("fitForDisabledWorkers param", () => {
-        const offer = new OfferEntityBuilder().build();
-
-        const notFitForDisabledWorkers = new EstablishmentAggregateBuilder()
-          .withEstablishmentSiret("00000000000001")
-          .withFitForDisabledWorkers("no")
-          .withLocationId(uuid())
-          .withOffers([offer])
-          .withUserRights([osefUserRight])
-          .build();
-
-        const fitForDisabledWorkers = new EstablishmentAggregateBuilder()
-          .withEstablishmentSiret("00000000000002")
-          .withFitForDisabledWorkers("yes-ft-certified")
-          .withLocationId(uuid())
-          .withOffers([offer])
-          .withUserRights([osefUserRight])
-          .build();
-
-        beforeEach(async () => {
-          await Promise.all(
-            [notFitForDisabledWorkers, fitForDisabledWorkers].map(
-              (establishment) =>
-                pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-                  establishment,
-                ),
-            ),
-          );
-        });
-
-        it("without param", async () => {
-          const results =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  sortedBy: "distance",
-                  distanceKm: 10,
-                  lat: fitForDisabledWorkers.establishment.locations[0].position
-                    .lat,
-                  lon: fitForDisabledWorkers.establishment.locations[0].position
-                    .lon,
-                },
-              },
-            );
-          expectToEqual(results, [
-            makeExpectedSearchResult({
-              establishment: notFitForDisabledWorkers,
-              withOffers: [offer],
-              withLocationAndDistance: {
-                ...notFitForDisabledWorkers.establishment.locations[0],
-                distance: 0,
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-            makeExpectedSearchResult({
-              establishment: fitForDisabledWorkers,
-              withOffers: [offer],
-              withLocationAndDistance: {
-                ...fitForDisabledWorkers.establishment.locations[0],
-                distance: 0,
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ]);
-        });
-
-        it("with param false", async () => {
-          const results =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  sortedBy: "distance",
-                  distanceKm: 10,
-                  lat: fitForDisabledWorkers.establishment.locations[0].position
-                    .lat,
-                  lon: fitForDisabledWorkers.establishment.locations[0].position
-                    .lon,
-                },
-                fitForDisabledWorkers: false,
-              },
-            );
-          expectToEqual(results, [
-            makeExpectedSearchResult({
-              establishment: notFitForDisabledWorkers,
-              withOffers: [offer],
-              withLocationAndDistance: {
-                ...notFitForDisabledWorkers.establishment.locations[0],
-                distance: 0,
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ]);
-        });
-
-        it("with param true", async () => {
-          const results =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  sortedBy: "distance",
-                  distanceKm: 10,
-                  lat: fitForDisabledWorkers.establishment.locations[0].position
-                    .lat,
-                  lon: fitForDisabledWorkers.establishment.locations[0].position
-                    .lon,
-                },
-                fitForDisabledWorkers: true,
-              },
-            );
-          expectToEqual(results, [
-            makeExpectedSearchResult({
-              establishment: fitForDisabledWorkers,
-              withOffers: [offer],
-              withLocationAndDistance: {
-                ...fitForDisabledWorkers.establishment.locations[0],
-                distance: 0,
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ]);
-        });
-      });
-
-      it("returns empty list when repo is empty", async () => {
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: cartographeSearchMade,
-            },
-          ),
-          [],
-        );
-      });
-
-      it("no location but keys provided in params - case occured from usecase without location", async () => {
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentWithOfferA1101_AtPosition,
-        );
-
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: {
-                sortedBy: "date",
-                voluntaryToImmersion: true,
-                place: undefined,
-                appellationCodes: undefined,
-                romeCodes: undefined,
-                searchableBy: "jobSeekers",
-                acquisitionCampaign: undefined,
-                acquisitionKeyword: undefined,
-              },
-            },
-          ),
-          [
-            makeExpectedSearchResult({
-              establishment: establishmentWithOfferA1101_AtPosition,
-              withOffers: [offer_A1101_11987],
-              withLocationAndDistance: {
-                ...locationOfSearchPosition,
-                distance: undefined,
-              },
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ],
-        );
-      });
-
-      it("returns open establishments only", async () => {
-        const openEstablishment = new EstablishmentAggregateBuilder()
-          .withEstablishmentSiret("00000000000001")
-          .withOffers([cartographeImmersionOffer])
-          .withLocations([locationOfSearchPosition])
-          .withEstablishmentOpen(true)
-          .withUserRights([osefUserRight])
-          .build();
-
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          openEstablishment,
-        );
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          closedEstablishment,
-        );
-
-        // Assert
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: cartographeSearchMade,
-            },
-          ),
-          [
-            makeExpectedSearchResult({
-              establishment: openEstablishment,
-              withLocationAndDistance: {
-                ...locationOfSearchPosition,
-                distance: 0,
-              },
-              withOffers: [cartographeImmersionOffer],
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ],
-        );
-      });
-
-      it("returns one search DTO by establishment, with offers matching rome and geographical area", async () => {
-        const establishmentAtRangeWithRome = new EstablishmentAggregateBuilder()
-          .withEstablishmentSiret("00000000000001")
-          .withOffers([
-            analysteEnGeomatiqueImmersionOffer,
-            cartographeImmersionOffer,
-          ])
-          .withLocations([locationOfSearchPosition])
-          .withUserRights([osefUserRight])
-          .build();
-
-        const establishmentAtRangeWithOtherRome =
-          new EstablishmentAggregateBuilder()
-            .withEstablishmentSiret("00000000000002")
-            .withOffers([artisteCirqueOffer])
-            .withLocations([locationOfCloseSearchPosition])
-            .withUserRights([osefUserRight])
-            .build();
-
-        const establishmentOutOfRangeWithRome =
-          new EstablishmentAggregateBuilder()
-            .withEstablishmentSiret("00000000000003")
-            .withOffers([
-              cartographeImmersionOffer,
-              analysteEnGeomatiqueImmersionOffer,
-            ])
-            .withLocations([locationOutOfAnySearchedPosition])
-            .withUserRights([osefUserRight])
-            .build();
-
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentAtRangeWithRome,
-        );
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentAtRangeWithOtherRome,
-        );
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentOutOfRangeWithRome,
-        );
-
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: cartographeSearchMade,
-            },
-          ),
-          [
-            makeExpectedSearchResult({
-              establishment: establishmentAtRangeWithRome,
-              withLocationAndDistance: {
-                ...locationOfSearchPosition,
-                distance: 0,
-              },
-              withOffers: establishmentAtRangeWithRome.offers,
-              nafLabel: "Activités des agences de travail temporaire",
-            }),
-          ],
-        );
-
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: { ...cartographeSearchMade, romeCodes: ["A1010"] },
-            },
-          ),
-          [],
-        );
-      });
-
-      describe("multiple locations", () => {
-        beforeEach(async () => {
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            establishmentCuvisteAtSaintesAndVeaux,
-          );
-
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            establishmentCuvisteAtChaniersAndLaRochelle,
-          );
-        });
-
-        it("search results with multiple locations and with matching rome code", async () => {
-          expectToEqual(
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  sortedBy: "date",
-                  distanceKm: 50,
-                  ...centerOfSaintesGeoPosition,
-                  romeCodes: [cuvisteOffer.romeCode],
-                },
-              },
-            ),
-            [
-              makeExpectedSearchResult({
-                establishment: establishmentCuvisteAtSaintesAndVeaux,
-                withLocationAndDistance: {
-                  ...bassompierreSaintesLocation,
-                  distance: 7704.55035665,
-                },
-                withOffers: [cuvisteOffer],
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-              makeExpectedSearchResult({
-                establishment: establishmentCuvisteAtChaniersAndLaRochelle,
-                withLocationAndDistance: {
-                  ...portHubleChaniersLocation,
-                  distance: 11093.36505388,
-                },
-                withOffers: [cuvisteOffer],
-                nafLabel: "Activités des agences de travail temporaire",
-              }),
-            ],
-          );
-        });
-
-        it("no search results with multiple locations and without matching rome code", async () => {
-          expectToEqual(
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: {
-                  sortedBy: "date",
-                  distanceKm: 100,
-                  ...centerOfSaintesGeoPosition,
-                  romeCodes: [groomChevauxOffer.romeCode],
-                },
-              },
-            ),
-            [],
-          );
-        });
-      });
-
-      describe("sorting", () => {
-        it("if sorted=distance, returns closest establishments in first", async () => {
-          const closeEstablishment = new EstablishmentAggregateBuilder()
-            .withEstablishmentSiret("00000000000001")
-            .withOffers([cartographeImmersionOffer])
-            .withLocations([
-              new LocationBuilder(locationOfSearchPosition)
-                .withId(uuid())
-                .build(),
-            ])
-            .withUserRights([osefUserRight])
-            .build();
-
-          const farEstablishment = new EstablishmentAggregateBuilder()
-            .withEstablishmentSiret("00000000000002")
-            .withOffers([cartographeImmersionOffer])
-            .withLocations([
-              new LocationBuilder(locationOfSearchPosition)
-
-                .withPosition({
-                  lat: locationOfSearchPosition.position.lat + 0.001,
-                  lon: locationOfSearchPosition.position.lon + 0.001,
-                })
-                .withId(uuid())
-                .build(),
-            ])
-            .withUserRights([osefUserRight])
-            .build();
-
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            closeEstablishment,
-          );
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            farEstablishment,
-          );
-
-          // Act
-          const searchResult =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: { ...cartographeSearchMade, sortedBy: "distance" },
-                maxResults: 2,
-              },
-            );
-
-          // Assert
-          expectToEqual(
-            searchResult.map((result) => result.siret),
-            [closeEstablishment, farEstablishment].map(
-              ({ establishment }) => establishment.siret,
-            ),
-          );
-        });
-
-        it("if sorted=date, returns latest offers in first", async () => {
-          const recentOfferEstablishment = new EstablishmentAggregateBuilder()
-            .withEstablishmentSiret("00000000000001")
-            .withOffers([
-              new OfferEntityBuilder(cartographeImmersionOffer)
-                .withCreatedAt(new Date("2022-05-05"))
-                .build(),
-            ])
-            .withLocations([
-              new LocationBuilder(locationOfSearchPosition)
-                .withId(uuid())
-                .build(),
-            ])
-            .withUserRights([osefUserRight])
-            .build();
-
-          const olderOfferEstablishment = new EstablishmentAggregateBuilder()
-            .withEstablishmentSiret("00000000000002")
-            .withOffers([
-              new OfferEntityBuilder(cartographeImmersionOffer)
-                .withCreatedAt(new Date("2022-05-02"))
-                .build(),
-            ])
-            .withLocations([
-              new LocationBuilder(locationOfSearchPosition)
-                .withId(uuid())
-                .build(),
-            ])
-            .withUserRights([osefUserRight])
-            .build();
-
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            recentOfferEstablishment,
-          );
-          await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-            olderOfferEstablishment,
-          );
-
-          // Act
-          const searchResult =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: { ...cartographeSearchMade, sortedBy: "date" },
-                maxResults: 2,
-              },
-            );
-
-          // Assert
-          expectToEqual(
-            searchResult.map((result) => result.siret),
-            [recentOfferEstablishment, olderOfferEstablishment].map(
-              ({ establishment }) => establishment.siret,
-            ),
-          );
-        });
-
-        it("if sorted=score, returns offers with appellations that have better score first", async () => {
-          const establishmentWithHighAndLowScore =
-            new EstablishmentAggregateBuilder()
-              .withEstablishmentSiret("00000000000001")
-              .withScore(10)
-              .withLocations([
-                {
-                  id: uuid(),
-                  position: locationOfSearchPosition.position,
-                  address: {
-                    city: "",
-                    departmentCode: "",
-                    postcode: "",
-                    streetNumberAndAddress: "",
-                  },
-                },
-              ])
-              .withOffers([
-                new OfferEntityBuilder(cartographeImmersionOffer).build(),
-                new OfferEntityBuilder(
-                  analysteEnGeomatiqueImmersionOffer,
-                ).build(),
-              ])
-              .withUserRights([osefUserRight])
-              .build();
-
-          const establishmentWithMediumScores =
-            new EstablishmentAggregateBuilder()
-              .withEstablishmentSiret("00000000000002")
-              .withScore(6)
-              .withLocations([
-                {
-                  id: uuid(),
-                  position: locationOfSearchPosition.position,
-                  address: {
-                    city: "",
-                    departmentCode: "",
-                    postcode: "",
-                    streetNumberAndAddress: "",
-                  },
-                },
-              ])
-              .withOffers([
-                new OfferEntityBuilder(cartographeImmersionOffer).build(),
-                new OfferEntityBuilder(
-                  analysteEnGeomatiqueImmersionOffer,
-                ).build(),
-              ])
-              .withUserRights([osefUserRight])
-              .build();
-
-          await Promise.all([
-            pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishmentWithHighAndLowScore,
-            ),
-            pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-              establishmentWithMediumScores,
-            ),
-          ]);
-
-          // Act
-          const searchResult =
-            await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-              {
-                searchMade: { ...cartographeSearchMade, sortedBy: "score" },
-              },
-            );
-          // Assert
-          expectToEqual(
-            searchResult.map(({ siret }) => siret),
-            [
-              establishmentWithHighAndLowScore,
-              establishmentWithMediumScores,
-            ].map(({ establishment }) => establishment.siret),
-          );
-        });
-      });
-
-      it("when multiple appellationCodes, returns the two related immersion-offers", async () => {
-        const mostRecentDate = new Date("2022-05-05");
-        const establishmentWithMostRecentOffer =
-          new EstablishmentAggregateBuilder()
-            .withEstablishmentUpdatedAt(mostRecentDate)
-            .withOffers([
-              new OfferEntityBuilder(cartographeImmersionOffer)
-                .withCreatedAt(mostRecentDate)
-                .build(),
-            ])
-            .withEstablishmentSiret("99000403200029")
-            .withLocations([
-              new LocationBuilder(locationOfSearchPosition)
-                .withId(uuid())
-                .build(),
-            ])
-            .withUserRights([osefUserRight])
-            .build();
-
-        const olderDate = new Date("2022-05-02");
-        const establishmentWithMostOlderOffer =
-          new EstablishmentAggregateBuilder()
-            .withEstablishmentUpdatedAt(olderDate)
-            .withOffers([
-              new OfferEntityBuilder(analysteEnGeomatiqueImmersionOffer)
-                .withCreatedAt(olderDate)
-                .build(),
-            ])
-            .withEstablishmentSiret("11000403200029")
-            .withLocations([
-              new LocationBuilder(locationOfSearchPosition)
-                .withId(uuid())
-                .build(),
-            ])
-            .withUserRights([osefUserRight])
-            .build();
-
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentWithMostRecentOffer,
-        );
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentWithMostOlderOffer,
-        );
-
-        // Act
-        const searchResult =
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: {
-                ...cartographeSearchMade,
-                sortedBy: "date",
-                appellationCodes: [
-                  cartographeImmersionOffer.appellationCode,
-                  analysteEnGeomatiqueImmersionOffer.appellationCode,
-                ],
-              },
-            },
-          );
-
-        // Assert
-        expectToEqual(
-          searchResult.map(({ siret }) => siret),
-          [
-            establishmentWithMostRecentOffer,
-            establishmentWithMostOlderOffer,
-          ].map(({ establishment }) => establishment.siret),
-        );
-      });
-
-      it("provide next availability date, update_date and created_at on result", async () => {
-        const recentDate = new Date("2023-12-02");
-        const aggregateWithOptionalValues = new EstablishmentAggregateBuilder()
-          .withLocations([
-            new LocationBuilder(portHubleChaniersLocation)
-              .withId(uuid())
-              .build(),
-          ])
-          .withOffers([
-            new OfferEntityBuilder().withCreatedAt(recentDate).build(),
-          ])
-          .withEstablishmentNextAvailabilityDate(new Date())
-          .withEstablishmentUpdatedAt(recentDate)
-          .withUserRights([osefUserRight])
-          .build();
-
-        const olderDate = new Date("2023-12-01");
-        const aggregateWithoutOptionalValues =
-          new EstablishmentAggregateBuilder()
-            .withLocations([
-              new LocationBuilder(portHubleChaniersLocation)
-                .withId(uuid())
-                .build(),
-            ])
-            .withOffers([
-              new OfferEntityBuilder().withCreatedAt(olderDate).build(),
-            ])
-            .withEstablishmentSiret("12341234123412")
-            .withEstablishmentNextAvailabilityDate(undefined)
-            .withEstablishmentUpdatedAt(olderDate)
-            .withUserRights([osefUserRight])
-            .build();
-
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          aggregateWithOptionalValues,
-        );
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          aggregateWithoutOptionalValues,
-        );
-
-        const searchResults =
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: {
-                ...aggregateWithOptionalValues.establishment.locations[0]
-                  .position,
-                appellationCodes: [
-                  aggregateWithOptionalValues.offers[0].appellationCode,
-                ],
-                distanceKm: 0,
-                sortedBy: "date",
-              },
-              maxResults: 2,
-            },
-          );
-
-        expectToEqual(
-          searchResults.map(
-            ({ siret, nextAvailabilityDate, updatedAt, createdAt }) => ({
-              siret,
-              nextAvailabilityDate,
-              updatedAt,
-              createdAt,
-            }),
-          ),
-          [
-            {
-              siret: aggregateWithOptionalValues.establishment.siret,
-              nextAvailabilityDate:
-                aggregateWithOptionalValues.establishment.nextAvailabilityDate,
-              updatedAt:
-                aggregateWithOptionalValues.establishment.updatedAt.toISOString(),
-              createdAt:
-                aggregateWithOptionalValues.establishment.createdAt.toISOString(),
-            },
-            {
-              siret: aggregateWithoutOptionalValues.establishment.siret,
-              nextAvailabilityDate: undefined,
-              updatedAt:
-                aggregateWithoutOptionalValues.establishment.updatedAt.toISOString(),
-              createdAt:
-                aggregateWithoutOptionalValues.establishment.createdAt.toISOString(),
-            },
-          ],
-        );
-      });
-      it("should return immersion offers even without lat/lon/distanceKm search", async () => {
-        const establishmentAggregate = new EstablishmentAggregateBuilder()
-          .withEstablishmentSiret("78000403200029")
-          .withOffers([cuvisteOffer, artisteCirqueOffer])
-          .withLocations([bassompierreSaintesLocation, veauxLocation])
-          .withUserRights([osefUserRight])
-          .build();
-
-        // Prepare
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentAggregate,
-        );
-
-        // Act
-        const searchResults: InternalOfferDto[] =
-          await pgEstablishmentAggregateRepository.legacySearchImmersionResults(
-            {
-              searchMade: {
-                sortedBy: "date",
-              },
-            },
-          );
-        const readableResults = searchResults.map(toReadableSearchResult);
-
-        // Assert
-        expect(readableResults).toHaveLength(4);
-      });
-
-      describe("Wrong paths", () => {
-        it("should throw on a search made with sortedBy distance and no geo params are provided", async () => {
-          // Assert
-          await expectPromiseToFailWithError(
-            pgEstablishmentAggregateRepository.legacySearchImmersionResults({
-              searchMade: {
-                sortedBy: "distance",
-              },
-            }),
-            errors.establishment.invalidGeoParams(),
-          );
-        });
-        it("should throw if only one of the geo params is provided (lat/lon/distanceKm)", async () => {
-          // Assert
-          await expectPromiseToFailWithError(
-            pgEstablishmentAggregateRepository.legacySearchImmersionResults({
-              searchMade: {
-                sortedBy: "distance",
-                lat: 45,
-              },
-            }),
-            errors.establishment.invalidGeoParams(),
-          );
-        });
-        it("should throw if all geo params value is 0 and sorted by distance", async () => {
-          // Assert
-          await expectPromiseToFailWithError(
-            pgEstablishmentAggregateRepository.legacySearchImmersionResults({
-              searchMade: {
-                sortedBy: "distance",
-                lat: 0,
-                lon: 0,
-                distanceKm: 0,
-              },
-            }),
-            errors.establishment.invalidGeoParams(),
-          );
-        });
-
-        it("should throw if lat / lon are 0 but distanceKm is provided and not 0 and sorted by distance", async () => {
-          // Assert
-          await expectPromiseToFailWithError(
-            pgEstablishmentAggregateRepository.legacySearchImmersionResults({
-              searchMade: {
-                sortedBy: "distance",
-                lat: 0,
-                lon: 0,
-                distanceKm: 10,
-              },
-            }),
-            errors.establishment.invalidGeoParams(),
-          );
-        });
-        it("should throw if one of the geo params value is 0", async () => {
-          // Assert
-          await expectPromiseToFailWithError(
-            pgEstablishmentAggregateRepository.legacySearchImmersionResults({
-              searchMade: {
-                sortedBy: "distance",
-                lat: 0,
-                lon: 45,
-                distanceKm: 10,
-              },
-            }),
-            errors.establishment.invalidGeoParams(),
-          );
-        });
-      });
-    });
 
     describe("getOffersAsAppellationDtoForFormEstablishment", () => {
       beforeEach(async () => {
@@ -3800,231 +2309,6 @@ describe("PgEstablishmentAggregateRepository", () => {
             romeCode: offer.romeCode,
             romeLabel: offer.romeLabel,
           })),
-        );
-      });
-    });
-
-    describe("getSearchResultBySearchQuery", () => {
-      beforeEach(async () => {
-        await pgEstablishmentAggregateRepository.insertEstablishmentAggregate(
-          establishmentWithOfferA1101_AtPositionWithTwoLocations,
-        );
-      });
-
-      it("undefined when missing offer by siret", async () => {
-        const missingSiret = "11111111111111";
-
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.getSearchResultBySearchQuery(
-            missingSiret,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-              .appellationCode,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .locations[0].id,
-          ),
-          undefined,
-        );
-      });
-
-      it("ignores location id when location does not exist anymore", async () => {
-        const deletedLocationId =
-          establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-            .locations[0].id;
-
-        await kyselyDb
-          .deleteFrom("establishments_location_infos")
-          .where("id", "=", deletedLocationId)
-          .execute();
-
-        const searchResultWithoutLocationId =
-          await pgEstablishmentAggregateRepository.getSearchResultBySearchQuery(
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .siret,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-              .appellationCode,
-          );
-
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.getSearchResultBySearchQuery(
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .siret,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-              .appellationCode,
-            deletedLocationId,
-          ),
-          searchResultWithoutLocationId,
-        );
-      });
-
-      it("undefined when missing offer appellation code", async () => {
-        const missingAppellationCode = artisteCirqueOffer.appellationCode;
-
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.getSearchResultBySearchQuery(
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .siret,
-            missingAppellationCode,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .locations[0].id,
-          ),
-          undefined,
-        );
-      });
-
-      it("Returns reconstructed SearchImmersionResultDto for given siret, appellationCode and given location id", async () => {
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.getSearchResultBySearchQuery(
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .siret,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-              .appellationCode,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .locations[0].id,
-          ),
-          {
-            rome: establishmentWithOfferA1101_AtPositionWithTwoLocations
-              .offers[0].romeCode,
-            romeLabel:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-                .romeLabel,
-            establishmentScore:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.score,
-            appellations: [
-              {
-                appellationLabel:
-                  establishmentWithOfferA1101_AtPositionWithTwoLocations
-                    .offers[0].appellationLabel,
-                appellationCode:
-                  establishmentWithOfferA1101_AtPositionWithTwoLocations
-                    .offers[0].appellationCode,
-              },
-            ],
-            naf: establishmentWithOfferA1101_AtPositionWithTwoLocations
-              .establishment.nafDto.code,
-            nafLabel: "Activités des agences de travail temporaire",
-            siret:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.siret,
-            name: establishmentWithOfferA1101_AtPositionWithTwoLocations
-              .establishment.name,
-            customizedName:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.customizedName,
-            website:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.website,
-            additionalInformation:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.additionalInformation,
-            voluntaryToImmersion:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.voluntaryToImmersion,
-            fitForDisabledWorkers:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.fitForDisabledWorkers,
-            numberOfEmployeeRange:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.numberEmployeesRange,
-            contactMode:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.contactMode,
-            distance_m: undefined,
-            address:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.locations[0].address,
-            position:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.locations[0].position,
-            locationId:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.locations[0].id,
-            createdAt:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment.createdAt.toISOString(),
-            updatedAt:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment.updatedAt.toISOString(),
-            remoteWorkMode:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-                .remoteWorkMode,
-            isAvailable: true,
-          },
-        );
-      });
-      it("Returns first reconstructed SearchImmersionResultDto for given siret, appellationCode but no location id", async () => {
-        expectToEqual(
-          await pgEstablishmentAggregateRepository.getSearchResultBySearchQuery(
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment
-              .siret,
-            establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-              .appellationCode,
-          ),
-          {
-            rome: establishmentWithOfferA1101_AtPositionWithTwoLocations
-              .offers[0].romeCode,
-            romeLabel:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-                .romeLabel,
-            establishmentScore:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.score,
-            appellations: [
-              {
-                appellationLabel:
-                  establishmentWithOfferA1101_AtPositionWithTwoLocations
-                    .offers[0].appellationLabel,
-                appellationCode:
-                  establishmentWithOfferA1101_AtPositionWithTwoLocations
-                    .offers[0].appellationCode,
-              },
-            ],
-            naf: establishmentWithOfferA1101_AtPositionWithTwoLocations
-              .establishment.nafDto.code,
-            nafLabel: "Activités des agences de travail temporaire",
-            siret:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.siret,
-            name: establishmentWithOfferA1101_AtPositionWithTwoLocations
-              .establishment.name,
-            customizedName:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.customizedName,
-            website:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.website,
-            additionalInformation:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.additionalInformation,
-            voluntaryToImmersion:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.voluntaryToImmersion,
-            fitForDisabledWorkers:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.fitForDisabledWorkers,
-            numberOfEmployeeRange:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.numberEmployeesRange,
-            contactMode:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.contactMode,
-            distance_m: undefined,
-            address:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.locations[1].address, // is the last location added
-            position:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.locations[1].position, // is the last location added
-            locationId:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations
-                .establishment.locations[1].id, // is the last location added
-            createdAt:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment.createdAt.toISOString(),
-            updatedAt:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.establishment.updatedAt.toISOString(),
-            remoteWorkMode:
-              establishmentWithOfferA1101_AtPositionWithTwoLocations.offers[0]
-                .remoteWorkMode,
-            isAvailable: true,
-          },
         );
       });
     });
