+++ back/src/domains/establishment/ports/EstablishmentAggregateQueries.ts
@@ -1,0 +1,10 @@
+import type { AppellationCode, LocationId, SiretDto } from "shared";
+import type { RepositorySearchResultDto } from "./EstablishmentAggregateRepository";
+
+export interface EstablishmentAggregateQueries {
+  getSearchResultBySearchQuery(
+    siret: SiretDto,
+    appellationCode: AppellationCode,
+    locationId?: LocationId,
+  ): Promise<RepositorySearchResultDto | undefined>;
+}
